+++ Engine/Engine/DGLCore.h
@@ -58,7 +58,7 @@ public:
 	float*normals;
 	float*tangents;
 	float*colors;
-	unsigned long*indices;
+	unsigned int*indices;
 } DGeometryBufferDesc;
 
 //抽象Geometry资源-用于实现不同API下的Geometry
@@ -67,7 +67,7 @@ class DGeometryRes
 public:
 	DGeometryRes(int vertexUsage /*顶点用法描述*/, bool dynamic /*是否为动态mesh*/);
 	void Refresh(DGeometryBufferDesc* desc); //更新顶点缓存
-	void Refresh(float* vertexbuffer, unsigned long* indexbuffer, int vertexCount, int indexCount);//更新顶点缓存
+	void Refresh(float* vertexbuffer, unsigned int* indexbuffer, int vertexCount, int indexCount);//更新顶点缓存
 	void DrawPrimitive(DGeometryTopology topology);//绘制
 	virtual void Release() = 0; //释放资源
 	bool IsInitialized();
@@ -75,8 +75,8 @@ public:
 protected:
 	//virtual void OnRefresh(DGeometryBufferDesc* desc) = 0;
 	//virtual bool OnInit(DGeometryBufferDesc* desc) = 0;
-	virtual void OnRefresh(float* vertexbuffer, unsigned long* indexbuffer, int vertexCount, int indexCount) = 0;
-	virtual bool OnInit(float* vertexbuffer, unsigned long* indexbuffer, int vertexCount, int indexCount) = 0;
+	virtual void OnRefresh(float* vertexbuffer, unsigned int* indexbuffer, int vertexCount, int indexCount) = 0;
+	virtual bool OnInit(float* vertexbuffer, unsigned int* indexbuffer, int vertexCount, int indexCount) = 0;
 	virtual void OnDraw(DGeometryTopology) = 0;
 
 protected:
