+++ Engine/Engine/DGeometryRes11.h
@@ -15,8 +15,8 @@ public:
 	virtual void Release();
 
 protected:
-	virtual void OnRefresh(float* vertexbuffer, unsigned long* indexbuffer, int vertexCount, int indexCount);
-	virtual bool OnInit(float* vertexbuffer, unsigned long* indexbuffer, int vertexCount, int indexCount);
+	virtual void OnRefresh(float* vertexbuffer, unsigned int* indexbuffer, int vertexCount, int indexCount);
+	virtual bool OnInit(float* vertexbuffer, unsigned int* indexbuffer, int vertexCount, int indexCount);
 	virtual void OnDraw(DGeometryTopology);
 
 private:
